renamed logging to h5 import module
